fix(effects): audit and fix FPS_ANIMATION_FACTOR misuse project-wide
Follow-up to the CCreditWin alpha-freeze fix -- swept skills/effects/
animation code for the same class of bug plus a related one, both triggered
by this branch's WIP SDL_GPU renderer running uncapped FPS on at least one
backend (Direct3D12 confirmed; Vulkan correctly vsync-caps and doesn't
trigger either).

Integer-accumulator truncation (freezes or wrong decay rate at high FPS,
same root cause as CCreditWin):
- ZzzEffectFireLeave.cpp: RainPosition (int -> float) -- froze the rain/
  leaf-fall scroll.
- UI/HUD/Notices.cpp: s_time (int -> float) -- decremented by a flat 1/call
  instead of the intended sub-1 delta, expiring notices too fast.

Double-application of FPS_ANIMATION_FACTOR (effect runs at factor^2 instead
of factor, i.e. far too sparse/slow):
- MoveHandlers.cpp: Move_BITMAP_FIRECRACKER's burst loop bound was scaled by
  the factor on top of CreateParticleFpsChecked()'s own per-call spawn
  chance (already the factor) -- burst got quadratically sparser at high FPS.
- ZzzEffectFireLeave.cpp: CreateLorenciaLeaf baked the factor into spawn
  velocity on top of MoveHeavenRain's own per-frame velocity scaling --
  leaves crawled instead of drifting.

Missing FPS_ANIMATION_FACTOR entirely (fixed per-call step with no time
basis -- runs N times too fast at high FPS instead of gliding/fading over
real time):
- DefaultCamera.cpp: UpdateCameraDistance's approach-to-target fraction,
  UpdateCustomCameraDistance's push-in/out step.
- MoveHandlers.cpp: Move_MODEL_DRAGON's tilt catch-up; Move_MODEL_
  PROTECTGUILD's spin/fade; six near-duplicate alpha/light fades sitting
  next to correctly-scaled sibling code (summoner head/summon effects, a
  casting effect, two Move_BITMAP_MAGIC subtypes, a flame-strike effect).
- GOBoid.cpp: Fenrir/Pegasus/Unicorn mount teleport fade-out.
- ZzzEffectFireLeave.cpp: MoveHeavenRain's tumbling rotation (Angle +=
  TurningForce) was missing the same dt scale its Velocity->Position
  sibling two lines above already uses correctly.

Not live-tested in-game yet -- flagging for verification, same as every
other fix in this audit chain.

+++ src/source/UI/HUD/Notices.cpp
@@ -23,7 +23,11 @@ namespace
     };
 
     int    s_count = 0;
-    int    s_time = NOTICE_LIFETIME;
+    // float, not int -- s_time -= FPS_ANIMATION_FACTOR below is a sub-1.0-per-call delta at high
+    // FPS, and int truncation was swallowing the whole delta and substituting a flat -1/call
+    // instead, defeating the FPS compensation FPS_ANIMATION_FACTOR exists to provide (same root
+    // cause as CCreditWin's alpha-fade freeze, opposite symptom: this decays too fast, not frozen).
+    float  s_time = NOTICE_LIFETIME;
     float  s_blinkPhase = 0.f;
     Notice s_notices[MAX_NOTICE];
